Account for relay prefix when sending messages
This allows for the relay (sender) prefix (AKA mask) that servers add to
messages. Without this, messages might be silently dropped for being too
long, even if they were short enough when transmitted.

We assume 100 chars for the messages will be enough, but the
`mask_length` argument can be passed in to control this if the true size
of the mask is known.