Make tests more informative
The separate tests are now colorfully delineated. Their success or
failure is also now explicit.

The warnings emitted during a test run are now clearly visible again,
and pytest now outputs a percentage (as is the default). An unfortunate
side-effect of upgrading pytest for the warnings output is that
pytest-greendots broke.